adds three-js driver
Got rid of rxjs-run since it was causing type errors with
@cycle/isolate, and it ended up being easier to convert between
xstream and rxjs manually where needed.

Added navbar which uses config from the three-js driver to figure out
what objects are supported. In order to add the navbar, the canvas had
to be added to the virtual DOM hierarchy, which mean that it couldn't
be initialized until the first DOM update. There's a bit of trickiness
to get the three.js driver initialized at the right time.

The UI controls are driven out of a schema which should probably be
pushed to the driver.

ignoring dist

works!

wip

dumped rxjs-run and got sliders working

controls are working

adds icosahedron

big reorg to support nav